Refactor object declaration move component to better place and move hooks and requests.
