+++ src/sources/components/Spectra/SpectraList.scss
