+++ src/sources/components/Spectra/SpectraModal.scss
