+++ src/sources/components/Thumbnail/Thumbnail.scss
